fix: cast logits to FP32 before cross_entropy + remove backward instrumentation
Two changes:

1. train.py: Cast logits to float32 before F.cross_entropy. BF16 softmax
   over vocab_size=129,280 causes catastrophic precision loss in the backward
   pass, producing NaN gradients in ALL parameters even though the forward
   pass (which uses the numerically stable log-sum-exp trick) produces
   finite logits and loss. The .float() ensures both forward softmax and
   backward gradient computation happen in FP32.

2. moe.py: Remove backward NaN instrumentation (confirmed MoE backward is
   clean — no NaN in dOut, Xe_pad, dequantized weights, H1/H3, dA, dH1/dH3,
   dX, or dGates). The .item() calls in the instrumentation caused GPU-CPU
   syncs that deadlocked with NCCL collectives during distributed backward.

+++ nmoe/train.py
@@ -251,7 +251,11 @@ def train(cfg: Config):
               logger.warning(f"[NAN] step={step_num} micro={micro_step}: logits contain NaN={nan_count} Inf={inf_count} (total={logits.numel()}, valid_max={logits_max:.4f})")
 
           with nvtx_ctx('train/loss'), time_ctx('time_ms/loss'):
-            loss_unreduced = F.cross_entropy(logits.reshape(-1, cfg.vocab_size), targets.reshape(-1), reduction='none')
+            # Cast logits to FP32 before cross_entropy: BF16 softmax over vocab_size=129,280
+            # causes catastrophic precision loss, producing NaN gradients in the backward pass.
+            # The .float() is intentionally placed before .reshape() so the full tensor is in FP32
+            # and both the forward softmax and backward gradient computation use FP32 precision.
+            loss_unreduced = F.cross_entropy(logits.float().reshape(-1, cfg.vocab_size), targets.reshape(-1), reduction='none')
             if loss_mask is not None:
               # SFT: use per-token loss mask from chat template (0=prompt, 1=response)
               mask = loss_mask.reshape(-1)
